Fix move distances and animation
Fix bug where not all 4 move distances were being used. Apparently I am terrible at modulo arithemtic.

The original moonstone kept the state of the current animation.

 If you moved left two frames, move right one frame and then moved left again. It would resume and play the third frame of the left animation.

I've removed the checks to reset the animation to frame 0 everytime input direction changes to match the original moonstone.

This is probably an uninteded quirk and arguably it would be better to start from zero each time. But let's stay faithful to the original.

(Should investigate whether this occured on the amiga verion)