Fix collision data structures.
The coordinates in collide.hit are just points instead of bounding
boxes, a collision box is created from the max x and y coord of the
bounding box.

Next a system should be added that checks for collision with the box and
then whether each pixel collides with the image data of the defender.