Fix game loop so update isn't called multiple times at the start.
If a move direction was held then the player would teleport across the
screen at the start of an encounter.

This is because the while loop was allowing multiple world.update calls,
but the draw calls hadn't occurred yet at start up. Locking down the
number of update calls allowed in one pass prevents this.

If this is run on a low end system such that the update calls can't keep
up then this might be an issue.