change specs word maintain to be at beginning of update.
Currently this is only used for dagger deletion.
If we have this after the dispatcher then the entity is deleted during
the collision frame which means it disappears a frame too early.
Switching it to before means it gets deleted a frame later so we see the
dagger connect with the target before disappearing.