append background name when saving image into cache.
This is mainly for the main font, the palette differs for the main
screen and for the select knight screen. to allow for this we add the
name of the background onto the end of the image name this way we cache
a different image for each type of background.

We could also base it off some hash of the paletee....