Animated map scene.
Three images are kept in a vec for the background images.
In the original this effect was achieved using palette swaps. We could
achieve the same effect using a shader that uses greyscale image and a
1d lookup texture and cycling the palettes using that.

This is left as a future TODO as this is the quickest way of
implementing an animated background at the sacrifice of some memory
usage.