Make Highlight player use with_thread_local
This means this system is run in time with the rendering loop, in the
DOS version of moonstone it would use the vertical blanking interval to
time the oscillating palette swaps and wait for two of these events
before updating the palettes.

If we have vsync enabled, this should work exactly the same way as the
original moonstone.