count bits instead of leading zeros to calculate bit planes.
The bits represents the bit planes to recombine. I can't remember why I
decide to minus of the leading zeros instead of just counting the ones.
This means that troggsp2.cel can actually be read. Before it was
interpretting the blit_type = 13 incorrectly as having 4 bit planes
instead of 3.

13 -> 0b00001101

Which has 3 ones ->  3 bit planes instead of 4.