Move terrain loading to be in EncounterScene::new()
previously I thought that the palette used by the scene would be one of
wab1.cmp, fob1.cmp, glb.cmp or swb.cmp. These files are the entire full
screen background used by lairs and encounters in that area/biome.

It turns out the palette in the "XXb1.cmp" are _only_ used for drawing
the background, and the palette used is actually from the "XX1.cmp"
files.

These files contain background scenery, trees, rocks, chests etc. This
palette is the one used to draw scenery and entities (knights, troggs).
It's just that so far, the practice scene is the wasteland, where both
the palete for wab1.cmp and wa1.cmp are the same, so I never noticed
before.

The palette setup and drawing in the dos version follows this order:

 - set the palette as the background palette __B1.cmp file.
 - draw the background
 - for each item of scenery:
   - set the palette as the __1.cmp file
   - draw each item of scenery
 - add palette swaps for the area/biome type
 - add palette swaps for enemies
 - add palette swaps for knights