Make Sprite derive LoadableYaml
Since Sprite is now transparent, it can use the LoadableYaml macro as
we're no longer constructing the Sprite wrapper around the animations.

I thought I wasn't a fan of LoadableYaml, mainly because of my lack of
understanding of macros, but this might be a better way than using the
GameRon wrapper and having to use game_ron.0 everywhere when loading
resources.