Check x step size in ConfirmVelocity before centering the x axis
We only want to center the x axis if the velocity is zero _and_ the
walking animation step size is zero.

Centering the x axis means that if the knight bumps into a wall or other
sprite then the animation is set as idle and the knight stops moving.

However, troggs have a 0 size velocity in one frame of the animation and
we don't want to replace that with the idle animation, so we check for
this special case before centering the x axis.